feat: OmokDao에 저장된 돌들 불러오는 기능 구현
