+++ app/src/main/java/woowacourse/omok/MainActivity.kt
@@ -7,11 +7,9 @@ import android.widget.TableRow
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.children
 import woowacourse.omok.model.board.Board
-import woowacourse.omok.model.db.OmokContract.OmokEntry
 import woowacourse.omok.model.db.OmokDAO
 import woowacourse.omok.model.db.OmokDbHelper
 import woowacourse.omok.model.entity.Point
-import woowacourse.omok.model.entity.Stone
 import woowacourse.omok.model.entity.StoneColor
 import woowacourse.omok.model.turn.BlackTurn
 import woowacourse.omok.model.turn.Finished
@@ -19,8 +17,6 @@ import woowacourse.omok.model.turn.Turn
 import woowacourse.omok.model.turn.WhiteTurn
 import woowacourse.omok.view.output.AndroidOutputView
 import woowacourse.omok.view.output.OutputView
-
-var currentTurn: Turn = BlackTurn(Board())
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -30,48 +26,18 @@ class MainActivity : AppCompatActivity() {
         val dbHelper = OmokDbHelper(this)
         val db = dbHelper.writableDatabase
         val dao = OmokDAO(db)
-
         val board = findViewById<TableLayout>(R.id.board)
         val outputView: OutputView = AndroidOutputView(board)
-
-        val cursor =
-            db.query(
-                OmokEntry.TABLE_NAME,
-                arrayOf(
-                    OmokEntry.POINT_X,
-                    OmokEntry.POINT_Y,
-                    OmokEntry.TURN,
-                ),
-                null,
-                null,
-                null,
-                null,
-                null,
-            )
-
-        val stones = mutableSetOf<Stone>()
-
-        while (cursor.moveToNext()) {
-            val x = cursor.getString(cursor.getColumnIndexOrThrow(OmokEntry.POINT_X))
-            val y = cursor.getString(cursor.getColumnIndexOrThrow(OmokEntry.POINT_Y))
-            val turn = cursor.getString(cursor.getColumnIndexOrThrow(OmokEntry.TURN))
-
-            stones.add(
-                Stone(
-                    Point(x.toInt(), y.toInt()),
-                    if (turn == "BLACK") StoneColor.BLACK else StoneColor.WHITE,
-                ),
-            )
-
-            currentTurn =
-                if (turn == "BLACK") {
-                    WhiteTurn(Board(stones))
-                } else {
-                    BlackTurn(Board(stones))
-                }
-        }
+        val stones = dao.selectAll()
+        var currentTurn: Turn =
+            when (stones.lastOrNull()?.stoneColor) {
+                StoneColor.BLACK -> WhiteTurn(Board(stones))
+                StoneColor.WHITE -> BlackTurn(Board(stones))
+                else -> BlackTurn(Board())
+            }
 
         printStartGuide(outputView)
+
         board.children
             .filterIsInstance<TableRow>()
             .flatMap { it.children }
